requirements.txt
flask
flask-restful
pandas
scikit-learn
python-dotenv